Move RegionManagerExtensions.RegisterViewWithRegion extension method to IRegionManager.
