+++ Source/Wpf/Prism.Wpf.Tests/Regions/RegionManagerFixture.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
+using Microsoft.Practices.ServiceLocation;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Prism.Regions;
 using Prism.Wpf.Tests.Mocks;
@@ -314,6 +315,81 @@ namespace Prism.Wpf.Tests.Regions
 
             Assert.IsTrue(regionManager.Regions["RegionName"].Views.Contains(view1));
             Assert.IsTrue(regionManager.Regions["RegionName"].Views.Contains(view2));
+        }
+
+        [TestMethod]
+        public void CanRegisterViewType()
+        {
+            try
+            {
+                var mockRegionContentRegistry = new MockRegionContentRegistry();
+
+                string regionName = null;
+                Type viewType = null;
+
+                mockRegionContentRegistry.RegisterContentWithViewType = (name, type) =>
+                {
+                    regionName = name;
+                    viewType = type;
+                    return null;
+                };
+                ServiceLocator.SetLocatorProvider(
+                    () => new MockServiceLocator
+                    {
+                        GetInstance = t => mockRegionContentRegistry
+                    });
+
+                var regionManager = new RegionManager();
+
+                regionManager.RegisterViewWithRegion("Region1", typeof(object));
+
+                Assert.AreEqual(regionName, "Region1");
+                Assert.AreEqual(viewType, typeof(object));
+
+
+            }
+            finally
+            {
+                ServiceLocator.SetLocatorProvider(null);
+            }
+        }
+
+        [TestMethod]
+        public void CanRegisterDelegate()
+        {
+            try
+            {
+                var mockRegionContentRegistry = new MockRegionContentRegistry();
+
+                string regionName = null;
+                Func<object> contentDelegate = null;
+
+                Func<object> expectedDelegate = () => true;
+                mockRegionContentRegistry.RegisterContentWithDelegate = (name, usedDelegate) =>
+                {
+                    regionName = name;
+                    contentDelegate = usedDelegate;
+                    return null;
+                };
+                ServiceLocator.SetLocatorProvider(
+                    () => new MockServiceLocator
+                    {
+                        GetInstance = t => mockRegionContentRegistry
+                    });
+
+                var regionManager = new RegionManager();
+
+                regionManager.RegisterViewWithRegion("Region1", expectedDelegate);
+
+                Assert.AreEqual("Region1", regionName);
+                Assert.AreEqual(expectedDelegate, contentDelegate);
+
+
+            }
+            finally
+            {
+                ServiceLocator.SetLocatorProvider(null);
+            }
         }
     }
