+++ Source/Wpf/Prism.Wpf/Regions/IRegionManager.cs
@@ -1,5 +1,6 @@
 
 
+using System;
 using System.Windows.Controls;
 namespace Prism.Regions
 {
@@ -26,5 +27,25 @@ namespace Prism.Regions
         /// <param name="view">The view to add to the views collection</param>
         /// <returns>The RegionManager, to easily add several views. </returns>
         IRegionManager AddToRegion(string regionName, object view);
+
+        /// <summary>
+        /// Associate a view with a region, by registering a type. When the region get's displayed
+        /// this type will be resolved using the ServiceLocator into a concrete instance. The instance
+        /// will be added to the Views collection of the region
+        /// </summary>
+        /// <param name="regionName">The name of the region to associate the view with.</param>
+        /// <param name="viewType">The type of the view to register with the </param>
+        /// <returns>The regionmanager, for adding several views easily</returns>
+        IRegionManager RegisterViewWithRegion(string regionName, Type viewType);
+
+        /// <summary>
+        /// Associate a view with a region, using a delegate to resolve a concreate instance of the view. 
+        /// When the region get's displayed, this delelgate will be called and the result will be added to the
+        /// views collection of the region. 
+        /// </summary>
+        /// <param name="regionName">The name of the region to associate the view with.</param>
+        /// <param name="getContentDelegate">The delegate used to resolve a concreate instance of the view.</param>
+        /// <returns>The regionmanager, for adding several views easily</returns>
+        IRegionManager RegisterViewWithRegion(string regionName, Func<object> getContentDelegate);
     }
 }
